pyproject.toml
[project]
name = "quiz-app-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "passlib>=1.7.4",
    "pydantic>=2.11.7",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
    "xlrd>=2.0.2",
]
